myEmitter workking in events.js file.
